feat: add v2 to sealed charts
