+++ components/one-data-chart.tsx
@@ -5,10 +5,12 @@ import * as echarts from 'echarts';
 import moment from 'moment';
 import Loader from '@/components/loader';
 import { chartColors } from '@/utils/colors';
-import { DataPreparedChartProps } from '@/utils/interfaces';
+import { DataPreparedChartProps } from '@utils/interfaces';
 
-const DataPreparedChart = ({ data, title }: DataPreparedChartProps) => {
+const OneDataChart = ({ data, title }: DataPreparedChartProps) => {
   const [chartHeight, setChartHeight] = useState('600px');
+  const xData = data && data[0] && data[0].data.map(entry => entry.x);
+  const yData = data && data[0] && data[0].data.map(entry => entry.y);
 
   useEffect(() => {
     const updateChartHeight = () => {
@@ -32,33 +34,11 @@ const DataPreparedChart = ({ data, title }: DataPreparedChartProps) => {
     };
   }, []);
 
-  const options = {
-    legend: {
-      orient: 'vertical',
-      top: 24,
-      left: 75,
-      data: data.map(item => ({
-        name: item.id,
-        icon: 'square',
-        textStyle: {
-          color: chartColors.axisLabelTextColor,
-          fontSize: 14,
-          fontFamily: 'SuisseIntl',
-          fontWeight: 500,
-          lineHeight: 20,
-          padding: [0, 0, 0, -4],
-        },
-      })),
-      padding: 15,
-      borderRadius: 5,
-      borderColor: chartColors.gridLineColor,
-      backgroundColor: chartColors.legendBg,
-      itemGap: 5,
-    },
+  const option = {
     tooltip: {
       trigger: 'axis',
-      backgroundColor: chartColors.tooltipBgGreen,
-      borderColor: chartColors.tooltipBgGreen,
+      backgroundColor: chartColors.tooltipBgOrange,
+      borderColor: chartColors.tooltipBgOrange,
       padding: [3, 10],
       textStyle: {
         fontFamily: 'SuisseIntl',
@@ -68,12 +48,9 @@ const DataPreparedChart = ({ data, title }: DataPreparedChartProps) => {
         lineHeight: 18
       },
       formatter: function (params:any) {
-        let tooltipText = '';
-        params.forEach((param:any) => {
-          tooltipText += param.marker + ' ' + param.seriesName + ': ' + byteSize(param.value[1]).toString() + '<br />';
-        });
-        tooltipText += moment(params[0].axisValue).format('MMM D, YYYY');
-        return tooltipText;
+        const xValue = params[0].name;
+        const yValue = byteSize(params[0].value).toString();
+        return `${yValue}<br/>${moment(xValue).format('MMM D, YYYY')}`;
       },
       axisPointer: {
         lineStyle: {
@@ -89,6 +66,7 @@ const DataPreparedChart = ({ data, title }: DataPreparedChartProps) => {
       containLabel: true
     },
     xAxis: {
+      data: xData,
       axisTick: false,
       type: 'category',
       splitLine: {
@@ -138,48 +116,44 @@ const DataPreparedChart = ({ data, title }: DataPreparedChartProps) => {
         },
       },
     },
-    series: data && data.map((item, index) => ({
-      data: item.data.map(d => ({
-        name: d.x,
-        value: [d.x, d.y],
-      })),
-      name: item.id,
+    series: [{
       type: 'line',
       smooth: true,
       areaStyle: {
         color: new echarts.graphic.LinearGradient(0, 0, 0, 1, [
           {
             offset: 0,
-            color: chartColors.green,
+            color: chartColors.orange,
           },
           {
             offset: 1,
-            color: chartColors.greenEnd,
+            color: chartColors.orangeEnd,
           },
         ]),
       },
       itemStyle: {
-        color: index === 0 ? chartColors.green : chartColors.orange,
+        color: chartColors.orange,
       },
       emphasis: {
         scale: 2.5,
         itemStyle: {
           color: chartColors.tooltipColor,
-          borderColor: index === 0 ? chartColors.green : chartColors.orange,
-        },
+          borderColor: chartColors.orange,
+        }
       },
+      data: yData,
       lineStyle: {
-        color: index === 0 ? chartColors.green : chartColors.orange,
+        color: chartColors.orange,
       },
-      showSymbol: false,
-    }))
+      showSymbol: false, // Hide symbols on the line
+    }],
   };
 
   return (
     <div className="col-6_md-12">
       <h2>{title}</h2>
-      {data && data.length > 0 ? (
-        <ReactECharts option={options} style={{ height: chartHeight, width: '100%' }} />
+      {data && data[0] && data[0].data.length > 0 ? (
+        <ReactECharts option={option} style={{ height: chartHeight, width: '100%' }} />
       ) : (
         <div className="chart-placeholder">
           <Loader />
@@ -189,4 +163,4 @@ const DataPreparedChart = ({ data, title }: DataPreparedChartProps) => {
   );
 };
 
-export default DataPreparedChart;
+export default OneDataChart;
