+++ components/two-data-chart.tsx
@@ -5,12 +5,10 @@ import * as echarts from 'echarts';
 import moment from 'moment';
 import Loader from '@/components/loader';
 import { chartColors } from '@/utils/colors';
-import { DataPreparedChartProps } from '@utils/interfaces';
+import { DataPreparedChartProps } from '@/utils/interfaces';
 
-const DealSealedChart = ({ data, title }: DataPreparedChartProps) => {
+const TwoDataChart = ({ data, title, theme }: DataPreparedChartProps) => {
   const [chartHeight, setChartHeight] = useState('600px');
-  const xData = data && data[0] && data[0].data.map(entry => entry.x);
-  const yData = data && data[0] && data[0].data.map(entry => entry.y);
 
   useEffect(() => {
     const updateChartHeight = () => {
@@ -34,11 +32,33 @@ const DealSealedChart = ({ data, title }: DataPreparedChartProps) => {
     };
   }, []);
 
-  const option = {
+  const options = {
+    legend: {
+      orient: 'vertical',
+      top: 24,
+      left: 75,
+      data: data.map(item => ({
+        name: item.id,
+        icon: 'square',
+        textStyle: {
+          color: chartColors.axisLabelTextColor,
+          fontSize: 14,
+          fontFamily: 'SuisseIntl',
+          fontWeight: 500,
+          lineHeight: 20,
+          padding: [0, 0, 0, -4],
+        },
+      })),
+      padding: 15,
+      borderRadius: 5,
+      borderColor: chartColors.gridLineColor,
+      backgroundColor: chartColors.legendBg,
+      itemGap: 5,
+    },
     tooltip: {
       trigger: 'axis',
-      backgroundColor: chartColors.tooltipBgOrange,
-      borderColor: chartColors.tooltipBgOrange,
+      backgroundColor: theme === 'green' ? chartColors.tooltipBgGreen : chartColors.tooltipBgOrange,
+      borderColor: theme === 'green' ? chartColors.tooltipBgGreen : chartColors.tooltipBgOrange,
       padding: [3, 10],
       textStyle: {
         fontFamily: 'SuisseIntl',
@@ -48,9 +68,12 @@ const DealSealedChart = ({ data, title }: DataPreparedChartProps) => {
         lineHeight: 18
       },
       formatter: function (params:any) {
-        const xValue = params[0].name;
-        const yValue = byteSize(params[0].value).toString();
-        return `${yValue}<br/>${moment(xValue).format('MMM D, YYYY')}`;
+        let tooltipText = '';
+        params.forEach((param:any) => {
+          tooltipText += param.marker + ' ' + param.seriesName + ': ' + byteSize(param.value[1]).toString() + '<br />';
+        });
+        tooltipText += moment(params[0].axisValue).format('MMM D, YYYY');
+        return tooltipText;
       },
       axisPointer: {
         lineStyle: {
@@ -66,7 +89,6 @@ const DealSealedChart = ({ data, title }: DataPreparedChartProps) => {
       containLabel: true
     },
     xAxis: {
-      data: xData,
       axisTick: false,
       type: 'category',
       splitLine: {
@@ -116,44 +138,48 @@ const DealSealedChart = ({ data, title }: DataPreparedChartProps) => {
         },
       },
     },
-    series: [{
+    series: data && data.map((item, index) => ({
+      data: item.data.map(d => ({
+        name: d.x,
+        value: [d.x, d.y],
+      })),
+      name: item.id,
       type: 'line',
       smooth: true,
       areaStyle: {
         color: new echarts.graphic.LinearGradient(0, 0, 0, 1, [
           {
             offset: 0,
-            color: chartColors.orange,
+            color: theme === 'green' ? chartColors.green : chartColors.orange,
           },
           {
             offset: 1,
-            color: chartColors.orangeEnd,
+            color: theme === 'green' ? chartColors.greenEnd : chartColors.orangeEnd,
           },
         ]),
       },
       itemStyle: {
-        color: chartColors.orange,
+        color: index === 0 ? theme === 'green' ? chartColors.green : chartColors.orange : theme === 'green' ? chartColors.orange : chartColors.green,
       },
       emphasis: {
         scale: 2.5,
         itemStyle: {
           color: chartColors.tooltipColor,
-          borderColor: chartColors.orange,
-        }
+          borderColor: index === 0 ? theme === 'green' ? chartColors.green : chartColors.orange : theme === 'green' ? chartColors.orange : chartColors.green,
+        },
       },
-      data: yData,
       lineStyle: {
-        color: chartColors.orange,
+        color: index === 0 ? theme === 'green' ? chartColors.green : chartColors.orange : theme === 'green' ? chartColors.orange : chartColors.green,
       },
-      showSymbol: false, // Hide symbols on the line
-    }],
+      showSymbol: false,
+    }))
   };
 
   return (
     <div className="col-6_md-12">
       <h2>{title}</h2>
-      {data && data[0] && data[0].data.length > 0 ? (
-        <ReactECharts option={option} style={{ height: chartHeight, width: '100%' }} />
+      {data && data.length > 0 ? (
+        <ReactECharts option={options} style={{ height: chartHeight, width: '100%' }} />
       ) : (
         <div className="chart-placeholder">
           <Loader />
@@ -163,4 +189,4 @@ const DealSealedChart = ({ data, title }: DataPreparedChartProps) => {
   );
 };
 
-export default DealSealedChart;
+export default TwoDataChart;
